Fix login session inconsistency after database redeployment
For the sake of security, the user_email_hash is added to the session.
When the database gets updated while users perform the activity,
ids in the sessions and in the database may mess up.
This means that a random user can access the data of another.
To prevent this, the session now has the user_email_hash
generated by generate_hash. This hash should be used
to confirm that a user with provided id has the same email stored
in the database

Resolves: GGRC-6367